requirements.txt
asn1crypto==0.24.0
awsebcli==3.12.1
blessed==1.14.2
botocore==1.8.26
cement==2.8.2
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
colorama==0.3.7
cryptography==2.1.4
Django==1.11
django-angular==2.0.0
django-common-helpers==0.9.1
django-cors-headers==2.1.0
django-cron==0.5.0
django-extensions==1.9.8
django-mongodb-engine==0.6.0
djangorestframework==3.7.3
djangotoolbox==1.8.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
ez-setup==0.9
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
mongoengine==0.9.0
mysqlclient==1.3.12
ndg-httpsclient==0.4.3
pathspec==0.5.0
pkg-resources==0.0.0
pyasn1==0.4.2
pycparser==2.18
pymongo==3.6.0
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
